# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Carve API
  description: |
    The Carve language-learning API. This document is the source of truth for
    every `/v1` endpoint; the TypeScript client in `apps/web/src/lib/api.ts`
    is regenerated from it and schemathesis runs property-based HTTP fuzz
    against it as part of the L4 testing layer.
  version: 0.3.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.carve.app
    description: Production
security: []  # default: no auth; protected operations explicitly set `security: [{ bearer: [] }]`
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    LanguageCode:
      type: string
      enum: [ja, zh-cn, zh-tw, zh, ko, en]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    TokenizeRequest:
      type: object
      required: [text, language]
      properties:
        text:     { type: string, maxLength: 50000 }
        language: { $ref: '#/components/schemas/LanguageCode' }
        mode:     { type: string, enum: [A, B, C], description: "Sudachi mode (JA only)" }
    TokenizeResponse:
      type: object
      required: [tokens]
      properties:
        tokens:
          type: array
          items: { type: object, additionalProperties: true }
    LookupRequest:
      type: object
      required: [surface, language]
      properties:
        surface:     { type: string, maxLength: 200 }
        language:    { $ref: '#/components/schemas/LanguageCode' }
        target_lang: { type: string, maxLength: 5, default: en }
        context:     { type: string, maxLength: 500 }
    LookupResponse:
      type: object
      properties:
        entries:
          type: array
          items: { type: object, additionalProperties: true }
    ScoreContentRequest:
      type: object
      required: [text, language]
      properties:
        text:     { type: string, maxLength: 200000 }
        language: { $ref: '#/components/schemas/LanguageCode' }
        known_lemmas:
          type: array
          items: { type: string }
    ScoreContentResponse:
      type: object
      properties:
        comprehension_pct: { type: number, minimum: 0, maximum: 1 }
        unknown_count:     { type: integer, minimum: 0 }
        known_count:       { type: integer, minimum: 0 }
        total_tokens:      { type: integer, minimum: 0 }
    TranslateRequest:
      type: object
      required: [text]
      properties:
        text:            { type: string, maxLength: 5000 }
        source_language: { type: string, default: ja }
        target_language: { $ref: '#/components/schemas/LanguageCode' }
    TranslateResponse:
      type: object
      properties:
        translation:     { type: string }
        source_language: { type: string }
        target_language: { type: string }

    # ── Export ────────────────────────────────────────────────────────────
...
paths:
...
  /v1/nlp/tokenize:
    post:
      tags: [nlp]
      operationId: tokenize
      summary: Tokenize text
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/TokenizeRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TokenizeResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: Unsupported language, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '502': { description: NLP service unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/nlp/lookup:
    post:
      tags: [nlp]
      operationId: lookup
      summary: Dictionary lookup
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/LookupRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/LookupResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '502': { description: NLP service unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/nlp/score-content:
    post:
      tags: [nlp]
      operationId: scoreContent
      summary: Score a chunk of text against the user's vocabulary
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ScoreContentRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ScoreContentResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '502': { description: NLP service unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/nlp/translate:
    post:
      tags: [nlp]
      operationId: translate
      summary: Translate a short snippet
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/TranslateRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TranslateResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '502': { description: NLP service unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /v1/nlp/select-sentence:
    post:
      tags: [nlp]
      operationId: selectMiningSentence
      summary: Rank i+1 mining sentence candidates
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: true } } }
      responses:
        '200': { description: Ranked sentence candidates }
        '401': { description: Unauthorized }
        '502': { description: NLP service unavailable }
  /v1/nlp/grammar/patterns:
    get:
      tags: [nlp]
      operationId: listGrammarPatterns
      summary: List supported grammar patterns
      security: [{ bearer: [] }]
      responses:
        '200': { description: Grammar catalog }
        '401': { description: Unauthorized }
...